tests/wpt/tests/content-security-policy/securitypolicyviolation/securitypolicyviolation-block-image-from-script.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/content-security-policy/support/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="img-src 'none'">
<body>
<script>
  async_test(t => {
    waitUntilEvent(window, "securitypolicyviolation")
      .then(t.step_func_done(e => {
        assert_equals(e.documentURI, document.location.toString());
        assert_equals(e.referrer, document.referrer);
        assert_equals(e.blockedURI, "http://{{domains[www]}}:{{ports[http][0]}}/content-security-policy/support/fail.png");
        assert_equals(e.violatedDirective, "img-src");
        assert_equals(e.effectiveDirective, "img-src");
        assert_equals(e.originalPolicy, "img-src \'none\'");
        assert_equals(e.disposition, "enforce");
        assert_equals(new URL(e.sourceFile).pathname, "/content-security-policy/support/inject-image.sub.js");
        // Per https://html.spec.whatwg.org/#relevant-mutations:
        //     The img or source HTML element insertion steps or HTML element removing steps count the mutation as a relevant mutation.
        // So when the src load is async, line 3 (appendChild, and thus the insertion steps) is what triggers the relevant load, not the src setter.
        assert_equals(e.lineNumber, 3);
        assert_equals(e.columnNumber, 15);
        assert_equals(e.statusCode, 200);
      }));

    var s = document.createElement("script");
    s.src = "/content-security-policy/support/inject-image.sub.js";
    document.body.appendChild(s);
  }, "Non-redirected cross-origin URLs are not stripped.");
</script>
