LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/securitypolicyviolation-block-image-from-script.sub-expected.txt


FAIL Non-redirected cross-origin URLs are not stripped. assert_equals: expected "http://127.0.0.1:8800/content-security-policy/support/fail.png" but got "http://127.0.0.1:8800"
